Rename `trait.rs` to `traits.rs` to avoid the use of reserved keyword syntax.
