+++ core/src/avm2/globals.rs
@@ -5,9 +5,9 @@ use crate::avm2::class::Class;
 use crate::avm2::method::NativeMethod;
 use crate::avm2::names::{Namespace, QName};
 use crate::avm2::object::{FunctionObject, Object, ScriptObject, TObject};
-use crate::avm2::r#trait::Trait;
 use crate::avm2::scope::Scope;
 use crate::avm2::string::AvmString;
+use crate::avm2::traits::Trait;
 use crate::avm2::value::Value;
 use crate::avm2::Error;
 use gc_arena::{Collect, GcCell, MutationContext};
