+++ core/src/avm2/object/primitive_object.rs
@@ -5,9 +5,9 @@ use crate::avm2::class::Class;
 use crate::avm2::names::{Namespace, QName};
 use crate::avm2::object::script_object::{ScriptObjectClass, ScriptObjectData};
 use crate::avm2::object::{Object, ObjectPtr, TObject};
-use crate::avm2::r#trait::Trait;
 use crate::avm2::scope::Scope;
 use crate::avm2::string::AvmString;
+use crate::avm2::traits::Trait;
 use crate::avm2::value::Value;
 use crate::avm2::Error;
 use crate::impl_avm2_custom_object;
