+++ core/src/avm2/traits.rs
